Move signup cli into client/cli/signup (#1337)
